Update to v1.0.3

+++ src/Gml.Web.Api.Domains/Integrations/AzuriomAuthResult.cs
@@ -18,7 +18,7 @@ public class AzuriomAuthResult
     public bool EmailVerified { get; set; }
 
     [JsonProperty("money")]
-    public int Money { get; set; }
+    public decimal Money { get; set; }
 
     [JsonProperty("banned")]
     public bool Banned { get; set; }
